+++ controllers/conversion.js
@@ -5,6 +5,7 @@ import Papa from "papaparse";
 import base64ToText from "../js/util/base64ToText.js";
 import { body, matchedData, validationResult } from "express-validator";
 import ExcelJS from "exceljs";
+import { convertJsonToYamlDomain } from "../js/convert/jsonToYamlDomain.js";
 
 const router = express.Router();
 
@@ -27,37 +28,12 @@ router.post("/json_to_yaml", (req, res) => {
 
 router.post("/json_to_yaml_domain", (req, res) => {
   try {
-    let convertedYaml = stringify(req.body, { lineWidth: 0 });
-    const lines = convertedYaml.split("\n");
-
-    const processedLines = lines.map((line) => {
-      const trimmedLine = line.trim();
-      if (trimmedLine.startsWith("text:") || trimmedLine.startsWith("- text:")) {
-        const index = line.indexOf(":");
-        const prefix = line.substring(0, index);
-        const value = line.substring(index + 1).trim();
-
-        if (value.startsWith("'") && value.endsWith("'")) {
-          const innerValue = value.slice(1, -1).replace(/"/g, '\\"');
-          return `${prefix}: "${innerValue}"`;
-        }
-
-        if (!value.startsWith('"') || !value.endsWith('"')) {
-          const escapedValue = value.replace(/"/g, '\\"');
-          return `${prefix}: "${escapedValue}"`;
-        }
-        return line;
-      }
-      return line;
-    });
-
-    convertedYaml = processedLines.join("\n");
+    const convertedYaml = convertJsonToYamlDomain(req.body);
     res.send({ json: convertedYaml });
   } catch (error) {
     res.status(500).json({ error: "Failed to create file", details: error.message });
   }
 });
-
 
 router.post("/json_to_yaml_data", (req, res) => {
   try {
